+++ apps/Conduit-UI/src/components/security/ClientsTab.tsx
@@ -9,6 +9,8 @@ import {
   TableContainer,
   TableHead,
   TableRow,
+  TableSortLabel,
+  Tooltip,
   useMediaQuery,
   useTheme,
 } from '@mui/material';
@@ -18,11 +20,18 @@ import DeleteIcon from '@mui/icons-material/Delete';
 import React, { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import ClientPlatformEnum, { IClient } from '../../models/security/SecurityModels';
-import { asyncDeleteClient, asyncGetAvailableClients } from '../../redux/slices/securitySlice';
+import {
+  asyncDeleteClient,
+  asyncGetAvailableClients,
+  clearClientSecret,
+} from '../../redux/slices/securitySlice';
 import { useAppSelector } from '../../redux/store';
 import CreateSecurityClientDialog from './CreateSecurityClientDialog';
-import { Add, Update } from '@mui/icons-material';
-import UpdateSecurityClientDialog from './UpdateSecurityClientDialog';
+import { Add, Edit, KeyboardArrowDown } from '@mui/icons-material';
+import UpdateSecurityClient from './UpdateSecurityClient';
+import { SideDrawerWrapper } from '@conduitplatform/ui-components';
+import ClientSecretDialog from './ClientSecretDialog';
+import { prepareSort } from '../../utils/prepareSort';
 
 const emptyClient = {
   _id: '',
@@ -42,11 +51,16 @@ const ClientsTab: React.FC = () => {
   const theme = useTheme();
   const smallScreen = useMediaQuery(theme.breakpoints.down('sm'));
   const [updateDialog, setUpdateDialog] = useState<boolean>(false);
+  const [secretDialog, setSecretDialog] = useState<boolean>(false);
+  const [sort, setSort] = useState<{ asc: boolean; index: string | null }>({
+    asc: false,
+    index: null,
+  });
   const [selectedClient, setSelectedClient] = useState<IClient>(emptyClient);
 
   useEffect(() => {
-    dispatch(asyncGetAvailableClients());
-  }, [dispatch]);
+    dispatch(asyncGetAvailableClients({ sort: prepareSort(sort) }));
+  }, [dispatch, sort]);
 
   const { availableClients } = useAppSelector((state) => state.securitySlice.data);
 
@@ -58,9 +72,22 @@ const ClientsTab: React.FC = () => {
     setOpenDialog(false);
   };
 
+  const handleSuccessfullClientCreation = () => {
+    setOpenDialog(false);
+
+    setTimeout(() => {
+      setSecretDialog(true);
+    }, 1500);
+  };
+
   const handleCloseUpdateDialog = () => {
     setUpdateDialog(false);
     setSelectedClient(emptyClient);
+  };
+
+  const handleCloseSecretDialog = () => {
+    setSecretDialog(false);
+    dispatch(clearClientSecret());
   };
 
   const handleOpenUpdateDialog = (client: IClient) => {
@@ -68,12 +95,37 @@ const ClientsTab: React.FC = () => {
     setUpdateDialog(true);
   };
 
+  const headCells = [
+    { label: 'Client ID', sort: 'clientId' },
+    { label: 'Alias', sort: 'alias' },
+    { label: 'Platform', sort: 'platform' },
+    { label: 'Domain', sort: 'domain' },
+    { label: 'Notes', sort: 'notes' },
+  ];
+
+  const onSelectedField = (index: string) => {
+    if (setSort !== undefined)
+      setSort((prevState: { asc: boolean; index: string | null }) => {
+        if (prevState.index === index) {
+          return { asc: !prevState.asc, index: index };
+        }
+        return { asc: prevState.asc, index: index };
+      });
+  };
+
+  const handleDirection = (dir: boolean) => {
+    if (dir) {
+      return 'asc';
+    }
+    return 'desc';
+  };
+
   return (
     <Container>
       <Paper sx={{ p: 4, borderRadius: 8 }}>
         <Box display="flex" justifyContent="space-between" alignItems="center">
           <Typography variant={'h6'}>Available Security Clients</Typography>
-          <Button variant="outlined" onClick={() => setOpenDialog(true)}>
+          <Button variant="contained" onClick={() => setOpenDialog(true)}>
             {<Add />}
             {smallScreen ? undefined : <Typography>Generate</Typography>}
           </Button>
@@ -82,21 +134,18 @@ const ClientsTab: React.FC = () => {
           <TableContainer sx={{ maxHeight: '69vh' }}>
             <Table stickyHeader>
               <TableHead>
-                <TableRow>
-                  <TableCell
-                    sx={{
-                      backgroundColor: 'background.paper',
-                    }}>
-                    Client ID
+                {headCells.map((headCell) => (
+                  <TableCell sx={{ backgroundColor: 'background.paper' }} key={headCell.sort}>
+                    <TableSortLabel
+                      IconComponent={KeyboardArrowDown}
+                      active={sort?.index === headCell.sort}
+                      direction={handleDirection(sort?.asc)}
+                      onClick={() => onSelectedField(headCell.sort)}>
+                      <Typography variant="body2">{headCell.label}</Typography>
+                    </TableSortLabel>
                   </TableCell>
-                  <TableCell sx={{ backgroundColor: 'background.paper' }}>Alias</TableCell>
-                  <TableCell sx={{ backgroundColor: 'background.paper' }}>Client Secret</TableCell>
-                  <TableCell sx={{ backgroundColor: 'background.paper' }}>Platform</TableCell>
-                  <TableCell sx={{ backgroundColor: 'background.paper' }}>Domain</TableCell>
-                  <TableCell sx={{ backgroundColor: 'background.paper' }} />
-                  <TableCell sx={{ backgroundColor: 'background.paper' }} />
-                  <TableCell sx={{ backgroundColor: 'background.paper' }} />
-                </TableRow>
+                ))}
+                <TableCell sx={{ backgroundColor: 'background.paper' }} />
               </TableHead>
               <TableBody>
                 {availableClients?.map((client: IClient, index: number) => (
@@ -109,13 +158,7 @@ const ClientsTab: React.FC = () => {
                       sx={{ maxWidth: '40px', overflow: 'hidden', textOverflow: 'ellipsis' }}>
                       <Typography variant={'caption'}>{client.alias || 'N/A'}</Typography>
                     </TableCell>
-                    <TableCell sx={{ maxWidth: '40px' }}>
-                      <Box>
-                        <span style={{ overflowWrap: 'break-word' }}>
-                          {client.clientSecret ? client.clientSecret : '****'}
-                        </span>
-                      </Box>
-                    </TableCell>
+
                     <TableCell sx={{ maxWidth: '20px' }}>
                       <Typography variant={'caption'}>{client.platform}</Typography>
                     </TableCell>
@@ -123,16 +166,28 @@ const ClientsTab: React.FC = () => {
                       sx={{ maxWidth: '40px', overflow: 'hidden', textOverflow: 'ellipsis' }}>
                       <Typography variant={'caption'}>{client.domain || 'N/A'}</Typography>
                     </TableCell>
-                    <TableCell sx={{ backgroundColor: 'background.paper' }} />
-                    <TableCell sx={{ backgroundColor: 'background.paper' }} />
+                    <TableCell
+                      sx={{
+                        whiteSpace: 'nowrap',
+                        maxWidth: '100px',
+                        overflow: 'hidden',
+                        textOverflow: 'ellipsis',
+                      }}>
+                      <Typography variant={'caption'}>{client.notes || 'N/A'}</Typography>
+                    </TableCell>
+
                     <TableCell>
-                      <Box display="flex" gap={1}>
-                        <IconButton onClick={() => handleDeletion(client._id)} size="large">
-                          <DeleteIcon color="error" />
-                        </IconButton>
-                        <IconButton onClick={() => handleOpenUpdateDialog(client)} size="large">
-                          <Update color="secondary" />
-                        </IconButton>
+                      <Box display="flex" justifyContent="flex-end" gap={1}>
+                        <Tooltip title="Delete security client">
+                          <IconButton onClick={() => handleDeletion(client._id)}>
+                            <DeleteIcon color="error" />
+                          </IconButton>
+                        </Tooltip>
+                        <Tooltip title="Edit security client">
+                          <IconButton onClick={() => handleOpenUpdateDialog(client)}>
+                            <Edit color="secondary" />
+                          </IconButton>
+                        </Tooltip>
                       </Box>
                     </TableCell>
                   </TableRow>
@@ -142,12 +197,23 @@ const ClientsTab: React.FC = () => {
           </TableContainer>
         </Box>
       </Paper>
-      <CreateSecurityClientDialog open={openDialog} handleClose={handleClose} />
-      <UpdateSecurityClientDialog
-        open={updateDialog}
-        handleClose={handleCloseUpdateDialog}
-        client={selectedClient}
+      <CreateSecurityClientDialog
+        open={openDialog}
+        handleClose={handleClose}
+        handleSuccess={handleSuccessfullClientCreation}
       />
+      <ClientSecretDialog open={secretDialog} handleClose={handleCloseSecretDialog} />
+      <SideDrawerWrapper
+        open={updateDialog}
+        title={`Edit client ${selectedClient._id}`}
+        closeDrawer={() => handleCloseUpdateDialog()}
+        width={750}>
+        <UpdateSecurityClient
+          availableClients={availableClients}
+          handleClose={handleCloseUpdateDialog}
+          client={selectedClient}
+        />
+      </SideDrawerWrapper>
     </Container>
   );
 };