+++ apps/Conduit-UI/src/redux/slices/storageSlice.ts
@@ -51,7 +51,7 @@ const initialState: IStorageSlice = {
   data: {
     config: {
       active: false,
-      provider: '',
+      provider: 'local',
       storagePath: '',
       allowContainerCreation: true,
       defaultContainer: '',
@@ -70,6 +70,11 @@ const initialState: IStorageSlice = {
         accountId: '',
         region: '',
         secretAccessKey: '',
+      },
+      aliyun: {
+        region: '',
+        accessKeyId: '',
+        accessKeySecret: '',
       },
     },
     containers: {